refactor: move borrow more modal to modal dir
